Move components to appropriate subdirectories
