+++ src/ggrc-client/js/components/assessment/info-pane/info-pane.js
@@ -12,7 +12,7 @@ import '../attach-button';
 import '../info-pane-save-status';
 import '../../comment/comment-add-form';
 import '../../comment/mapped-comments';
-import '../mapped-objects/mapped-controls';
+import '../mapped-controls/mapped-controls';
 import '../../assessment/map-button-using-assessment-type';
 import '../../ca-object/ca-object-modal-content';
 import '../../comment/comment-add-form';
