+++ src/ggrc-client/js/components/assessment/mapped-control-related-objects/mapped-control-related-objects.stache
