+++ src/ggrc-client/js/components/assessment/mapped-controls/mapped-controls.stache
